+++ src/components/Playerify.vue
@@ -8,24 +8,39 @@ import { computed, reactive, Ref, ref, shallowRef, useTemplateRef, watch } from 
 import { PlayerType } from '@playerify/enums'
 import { PlaylistItem, PlaylistItemInternal } from '@playerify/types'
 import { isFloat, stringify, processPlaylist } from '@playerify/utils'
+import { useListen } from '@playerify/composables/useEventBus'
 
 const props = defineProps({
   type: { validator: (value: PlayerType) => Object.values(PlayerType).includes(value) },
   coverImage: { type: String },
-  showPlaylist: { type: Boolean, default: false },
+  coverImageOriginalSize: { type: Boolean, default: true },
   playlist: { type: Array<string|PlaylistItem>, default: () => [] },
   playlistVariant: { type: String, default: 'elevated' },
   playlistButtonColor: { type: String, default: 'default' },
   frameWidth: { type: String, default: '100%' },
   frameHeight: { type: String, default: '' },
+  hideFileName: { type: Boolean, default: false },
+  hideDuration: { type: Boolean, default: false },
+  hideAllControls: { type: Boolean, default: false },
+  hideProgress: { type: Boolean, default: false },
+  hideMainControls: { type: Boolean, default: false },
+  hideExtraControls: { type: Boolean, default: false },
+  hidePlayButton: { type: Boolean, default: false },
+  hidePrevTrackButton: { type: Boolean, default: false },
+  hideNextTrackButton: { type: Boolean, default: false },
+  hideVolumeButton: { type: Boolean, default: false },
+  hidePlaylistButton: { type: Boolean, default: false },
+  hidePlaybackRateButton: { type: Boolean, default: false },
+  hideExtrasButton: { type: Boolean, default: false },
+  hideFullscreenButton: { type: Boolean, default: false },
   playButtonColor: { type: String, default: 'default' },
   pauseButtonColor: { type: String, default: 'default' },
   volumeButtonColor: { type: String, default: 'default' },
   volumeOffButtonColor: { type: String, default: 'default' },
   playbackRateButtonColor: { type: String, default: 'default' },
-  settingsButtonColor: { type: String, default: 'default' },
+  extrasButtonColor: { type: String, default: 'default' },
   fullscreenButtonColor: { type: String, default: 'default' },
-  btnRounded: { type: String, default: 'sm' },
+  btnRounded: { type: String, default: 'md' },
   progressRounded: { type: String, default: 'sm' },
   progressSliderColor: { type: String, default: 'primary' },
   volumeSliderColor: { type: String, default: 'primary' },
@@ -33,13 +48,11 @@ const props = defineProps({
   defaultVolume: { type: Number, default: 0.8 },
   progressColor: { type: String, default: 'primary' },
   bufferColor: { type: String, default: 'secondary' },
-  showFileName: { type: Boolean, default: true },
-  showDuration: { type: Boolean, default: true },
   permanentVolumeSlider: { type: Boolean, default: true },
   debug: { type: Boolean, default: false },
 })
 
-const openPlaylist = ref(props.showPlaylist)
+const openPlaylist = ref(!props.hidePlaylistButton)
 
 const defaultType = ref(props.type)
 
@@ -50,10 +63,10 @@ const currentCoverImage = ref(props.coverImage)
 
 const coverImageAttributes = computed(() => {
   const attributes: { width: string, height?: string } = { width: 'auto' }
-  if (props.frameWidth) {
+  if (props.frameWidth && !props.coverImageOriginalSize) {
     attributes.width = props.frameWidth
   }
-  if (props.frameHeight) {
+  if (props.frameHeight && !props.coverImageOriginalSize) {
     attributes.height = props.frameHeight
   }
   return attributes
@@ -120,7 +133,7 @@ const endBuffer = computed(() => {
       : 0
 })
 
-controls.value.error = 'No src or playlist provided'
+controls.value.error = 'No playlist provided'
 
 watch(() => currentMedia.value, () => {
   if (media.value !== null) {
@@ -138,6 +151,26 @@ watch(() => props.playlist, () => {
   }
 }, { deep: true, immediate: true })
 
+const trackIndex = computed(() => playList.value.findIndex(item => item.src === currentMedia.value?.src))
+const prevTrackAvailable = computed(() => {
+  return trackIndex.value > 0
+})
+
+const nextTrackAvailable = computed(() => {
+  return trackIndex.value < (playList.value.length - 1)
+})
+
+useListen('playerify--prev-track', () => {
+  if (prevTrackAvailable.value) {
+    currentMedia.value = playList.value[trackIndex.value - 1]
+  }
+})
+useListen('playerify--next-track', () => {
+  if (nextTrackAvailable.value) {
+    currentMedia.value = playList.value[trackIndex.value + 1]
+  }
+})
+
 function toggleFullscreen () {
   if (currentType.value !== PlayerType.VIDEO) {
     return
@@ -148,6 +181,14 @@ function toggleFullscreen () {
   } else {
     document.exitFullscreen?.()
   }
+}
+
+function togglePictureInPicture () {
+  if (currentType.value !== PlayerType.VIDEO) {
+    return
+  }
+
+  playerControls.value.togglePictureInPicture()
 }
 
 watch(() => currentMedia.value, () => {
@@ -174,7 +215,7 @@ watch(() => props.coverImage, () => {
         </div>
         <div v-else class="mt-5 relative rounded-md shadow overflow-hidden text-center">
           <template v-if="currentType === PlayerType.AUDIO">
-            <div v-if="currentCoverImage" class="cover-image">
+            <div v-if="currentCoverImage" :class="coverImageOriginalSize ? 'cover-image' : ''">
               <VImg :src="currentCoverImage" v-bind="coverImageAttributes" class="ma-auto" />
             </div>
             <audio
@@ -197,7 +238,7 @@ watch(() => props.coverImage, () => {
               @click="playerControls.playing = !playerControls.playing"
             />
           </template>
-          <div v-if="showFileName">
+          <div v-if="!hideFileName" :class="currentType === PlayerType.VIDEO ? 'mt-4' : ''">
             {{ currentFileName }}
           </div>
           <!--      <div>-->
@@ -205,7 +246,7 @@ watch(() => props.coverImage, () => {
           <!--      </div>-->
         </div>
 
-        <div class="w-100 d-flex align-center">
+        <div v-if="!hideAllControls && !hideProgress" class="w-100 d-flex align-center">
           <Progress
             v-model:current-time="playerControls.currentTime"
             :disabled="media === null"
@@ -216,16 +257,23 @@ watch(() => props.coverImage, () => {
             :progress-slider-color="progressSliderColor"
             :progress-color="progressColor"
             :buffer-color="bufferColor"
-            :show-duration="showDuration"
+            :hide-duration="hideDuration"
           />
         </div>
 
-        <div class="d-flex flex-row">
+        <div v-if="!hideAllControls" class="d-flex flex-row">
           <BaseControls
+            v-if="!hideMainControls"
             v-model:playing="playerControls.playing"
             v-model:volume="playerControls.volume"
             v-model:muted="playerControls.muted"
             :disabled="media === null"
+            :disabled-prev-track-button="!prevTrackAvailable"
+            :disabled-next-track-button="!nextTrackAvailable"
+            :hide-prev-track-button="hidePrevTrackButton"
+            :hide-next-track-button="hideNextTrackButton"
+            :hide-play-button="hidePlayButton"
+            :hide-volume-button="hideVolumeButton"
             :btn-rounded="btnRounded"
             :play-button-color="playButtonColor"
             :pause-button-color="pauseButtonColor"
@@ -235,25 +283,31 @@ watch(() => props.coverImage, () => {
             :permanent-volume-slider="permanentVolumeSlider"
           />
 
-          <div class="ml-auto extra-actions">
+          <div v-if="!hideExtraControls" class="ml-auto extra-actions">
             <ExtraControls
               v-model:media="media"
               v-model:loop="loop"
+              v-model:is-picture-in-picture="playerControls.isPictureInPicture"
               v-model:open-playlist="openPlaylist"
               :disabled="media === null"
+              :hide-playlist-button="hidePlaylistButton"
+              :hide-playback-rate-button="hidePlaybackRateButton"
+              :hide-extras-button="hideExtrasButton"
+              :hide-fullscreen-button="hideFullscreenButton"
+              :supports-picture-in-picture="playerControls.supportsPictureInPicture"
               :btn-rounded="btnRounded"
-              :show-playlist="showPlaylist"
               :playlist-button-color="playlistButtonColor"
               :playback-rate-button-color="playbackRateButtonColor"
-              :settings-button-color="settingsButtonColor"
+              :extras-button-color="extrasButtonColor"
               :fullscreen-button-color="fullscreenButtonColor"
               :is-video="currentMedia?.type === PlayerType.VIDEO"
               @toggle-fullscreen="toggleFullscreen"
+              @toggle-picture-in-picture="togglePictureInPicture"
             />
           </div>
         </div>
 
-        <div v-if="showPlaylist" class="mt-3">
+        <div v-if="!hideAllControls && !hideExtraControls" class="mt-3">
           <Playlist v-model:media="currentMedia" v-model:open-playlist="openPlaylist" :playlist="playList" :playlist-variant="playlistVariant" />
         </div>
 
@@ -264,14 +318,12 @@ watch(() => props.coverImage, () => {
 </template>
 
 <style lang="scss" scoped>
-.extra-actions {
-  width: 140px;
-  display: flex;
-  justify-content: flex-end;
+.cover-image:deep(.v-img img) {
+  object-fit: none;
 }
 
-.volume-slider {
-  width: calc(100% - 50px);
-  max-width: 300px;
+.extra-actions {
+  display: flex;
+  justify-content: flex-end;
 }
 </style>