+++ src/components/Progress.vue
@@ -12,22 +12,22 @@ defineProps({
   progressSliderColor: { type: String, default: 'primary' },
   progressColor: { type: String, default: 'primary' },
   bufferColor: { type: String, default: 'secondary' },
-  showDuration: { type: Boolean, default: true },
+  hideDuration: { type: Boolean, default: false },
 })
 </script>
 
 <template>
   <div class="slider-buffer-container">
     <VSlider
-        v-model="currentTime"
-        :disabled="disabled"
-        :color="progressSliderColor"
-        min="0"
-        :max="duration"
-        :hide-details="true"
-        track-color="transparent"
-        :thumb-label="showDuration ? false : 'always'"
-        class="play-slider"
+      v-model="currentTime"
+      :disabled="disabled"
+      :color="progressSliderColor"
+      min="0"
+      :max="duration"
+      :hide-details="true"
+      track-color="transparent"
+      :thumb-label="hideDuration ? 'always' : false"
+      class="play-slider"
     >
       <template #thumb-label="{ modelValue }">
         {{ formatDuration(!waiting ? modelValue : null) }}
@@ -35,20 +35,20 @@ defineProps({
     </VSlider>
 
     <VProgressLinear
-        v-if="!disabled"
-        v-model="currentTime"
-        :buffer-value="endBuffer"
-        :color="progressColor"
-        :buffer-color="bufferColor"
-        :max="duration"
-        :indeterminate="waiting"
-        :rounded="progressRounded"
-        height="6"
-        class="buffer-bar"
+      v-if="!disabled"
+      v-model="currentTime"
+      :buffer-value="endBuffer"
+      :color="progressColor"
+      :buffer-color="bufferColor"
+      :max="duration"
+      :indeterminate="waiting"
+      :rounded="progressRounded"
+      height="6"
+      class="buffer-bar"
     />
   </div>
 
-  <span v-if="showDuration" class="media-progress-duration">
+  <span v-if="!hideDuration" class="media-progress-duration">
     {{ formatDuration(!waiting ? currentTime : null) }} / {{ formatDuration(!waiting ? duration : null) }}
   </span>
 </template>